requirements.txt
astropy == 5.2.2
numpy == 1.24.4
pyerfa == 2.0.0.3